caly dzien walki tylko dlatego ze ktos napisal test tak ze daje wartosci ktore robia reset nawet zmiennej long. Dlaczego spoj nie da danych wejsciowych ktore wywolaly blad? glupota. Caly dzien siedzenia i rozkminiania bo trzeba bylo zastosowac long long.....

+++ FZI_STEF/FZI_STEF.cpp
@@ -7,11 +7,12 @@ int main()
 {
     int cities;
     int plusNumber = 0;
-    int highestNum = 0;
-    int sum = 0;
+    long long highestNum = 0;
+    long long sum = 0;
     int* money;
     cin >> cities;
     money = new int[cities];
+
     for (int i = 0; i < cities; i++) {
         cin >> money[i];
         if (money[i] > 0) {
